Added styles
