--- src/components/Drawer/CustomDrawer.styles.ts
@@ -1,14 +1,0 @@
-import {StyleSheet} from 'react-native'
-
-const styles = StyleSheet.create({
-  titleBox: {margin: 20},
-  title: {
-    color: 'black',
-    textAlign: 'center',
-    fontWeight: 'bold',
-    fontSize: 20,
-  },
-  labelStyle: {color: 'black', fontSize: 20},
-})
-
-export {styles}
