+++ src/components/Drawer/CustomDrawer.styles.tsx
@@ -1,0 +1,43 @@
+import {StyleSheet} from 'react-native'
+
+const styles = StyleSheet.create({
+  titleBox: {margin: 20},
+  title: {
+    color: 'black',
+    fontWeight: 'bold',
+    fontSize: 20,
+  },
+  labelStyle: {
+    color: 'black',
+    fontSize: 20,
+  },
+  drawerSection: {
+    borderTopWidth: 1,
+  },
+  drawerSections: {
+    display: 'flex',
+    flex: 1,
+    flexDirection: 'column',
+    alignContent: 'space-between',
+    justifyContent: 'space-between',
+  },
+  drawerItem: {
+    display: 'flex',
+    marginVertical: 0,
+  },
+  face: {
+    margin: 0,
+    padding: 0,
+    transform: [{rotate: '90deg'}],
+    width: 32,
+    fontSize: 24,
+    textAlign: 'center',
+  },
+  faceBorder: {
+    borderWidth: 1,
+    borderRadius: 50,
+    width: 32,
+  },
+})
+
+export {styles}
